Change unwrap/1 to only return the value
When it returns a tagged tuple, it is much harder to use it inline. And
there isn't much benefit to having it return a tagged tuple, because if
the argument is the correct type, then it will unwrap successfully, and
if it's not, the unwrap function will raise a FunctionClauseError.
Dialyzer should also complain if you try to pass in the wrong type.