Add `new!/1` that returns value directly or raises
Mainly for use in test scripts when you need a value of the given type
and are under full control of the base value being used. In such cases,
this function avoids the boilerplate of deconstructing a tagged tuple as
returned from `new/1`. It is NOT recommended to use this in production
code, because exceptions should never be used as flow-control
mechanisms.