Make typespec non-opaque for newtypes
While I would much rather have these be opaque types, doing so makes it
really difficult to rely on Dialyzer checking, because Dialyzer is
unable to reconcile the literal structs it sees being passed around
against the opaque types.

I've also removed the `is_type` guard, because it is pointless in light
of no longer using opaque types. It is now more idiomatic to either
pattern-match on the struct type or use the `is_struct/2` guard clause.