Ecto type should be able to load invalid data
If data in a database is invalid, well...it was already stored, so we still need to handle it somehow. Don't error out when loading it, but also log a
warning if it is invalid.

It is necessary to use %Struct{key: value} style initialization so that we don't call any validation logic, etc. that is called by `new/1`.